+++ example/types.check.tsx
@@ -8,7 +8,10 @@ import type {
 } from '@react-navigation/drawer';
 import type {
   CompositeScreenProps,
+  NavigationHelpers,
   NavigatorScreenParams,
+  RouteConfigComponent,
+  RouteProp,
 } from '@react-navigation/native';
 import type { StackNavigationOptions } from '@react-navigation/stack';
 import {
@@ -16,6 +19,7 @@ import {
   StackScreenProps,
 } from '@react-navigation/stack';
 import { expectTypeOf } from 'expect-type';
+import React, { FC } from 'react';
 
 type RootStackParamList = {
   Home: NavigatorScreenParams<HomeDrawerParamList>;
@@ -211,3 +215,99 @@ export const LatestScreen = ({
     .parameter(0)
     .toEqualTypeOf<'LeftDrawer' | 'BottomTabs' | undefined>();
 };
+
+// ================================================
+// The checks below uses jest syntax for semantics only.
+// These tests will never actually run.
+// Notice the use of // @ts-expect-error for assertions
+// instead of `expectTypeOf`. This is because `expect-type` fails
+// to throw errors in various cases involving parameters or union types.
+// ================================================
+
+describe('RouteConfigComponent', () => {
+  type ParamList = {
+    hasParam: { param: string };
+    hasParam2: { param2: string };
+    noParam: undefined;
+  };
+
+  const Screen = <Name extends keyof ParamList>(
+    _: { name: Name } & RouteConfigComponent<ParamList, Name>
+  ) => null;
+
+  it("doesn't accept incorrect route params", () => {
+    const Component: FC<{ route: RouteProp<ParamList, 'hasParam'> }> = () =>
+      null;
+    // @ts-expect-error
+    <Screen name="hasParam2" component={Component} />;
+    // @ts-expect-error
+    <Screen name="noParam" component={Component} />;
+    // ok
+    <Screen name="hasParam" component={Component} />;
+  });
+
+  it("doesn't require the component to accept the `route` or `navigation` prop", () => {
+    const Component: FC<{}> = () => null;
+    // ok
+    <Screen name="hasParam" component={Component} />;
+    // ok
+    <Screen name="noParam" component={Component} />;
+  });
+
+  it('allows the component to declare any optional props', () => {
+    const Component: FC<{ someProp?: string }> = () => null;
+    <Screen name="hasParam" component={Component} />;
+    <Screen name="noParam" component={Component} />;
+  });
+
+  it("doesn't allow a required prop that's neither `route` nor `navigation`", () => {
+    const Component: FC<{ someProp: string }> = () => null;
+    // @ts-expect-error
+    <Screen name="hasParam" component={Component} />;
+    // @ts-expect-error
+    <Screen name="noParam" component={Component} />;
+  });
+
+  it('allows the component to accept just the `navigation` prop', () => {
+    const Component: FC<{ navigation: object }> = () => null;
+    // ok
+    <Screen name="hasParam" component={Component} />;
+    // ok
+    <Screen name="noParam" component={Component} />;
+  });
+});
+
+describe('NavigationHelpers.navigate', () => {
+  type ParamList = {
+    hasParam: { param: string };
+    hasParam2: { param2: string };
+    noParam: undefined;
+  };
+  const navigate: NavigationHelpers<ParamList>['navigate'] = () => {};
+
+  it('strictly checks type of route params', () => {
+    // ok
+    navigate('noParam');
+    // ok
+    navigate('hasParam', { param: '123' });
+    // @ts-expect-error
+    navigate('hasParam2', { param: '123' });
+  });
+
+  it('strictly checks type of route params when a union RouteName is passed', () => {
+    let routeName = undefined as unknown as keyof ParamList;
+
+    // @ts-expect-error
+    navigate(routeName);
+
+    // ok
+    if (routeName === 'noParam') navigate(routeName);
+    // ok
+    if (routeName === 'hasParam') navigate(routeName, { param: '123' });
+
+    // @ts-expect-error
+    if (routeName === 'hasParam') navigate(routeName);
+    // @ts-expect-error
+    if (routeName === 'hasParam2') navigate(routeName, { param: '123' });
+  });
+});
